+++ custom_components/narwal_cloud/config_flow.py
@@ -6,7 +6,6 @@ import uuid
 from typing import Any
 
 import voluptuous as vol
-
 from homeassistant import config_entries
 from homeassistant.core import callback
 from homeassistant.helpers import selector